+++ app/controllers/frontend/characters/feats_controller.rb
@@ -37,6 +37,7 @@ module Frontend
         when 'pathfinder2' then authorized_scope(Character.all).pathfinder2
         when 'daggerheart' then authorized_scope(Character.all).daggerheart
         when 'dc20' then authorized_scope(Character.all).dc20
+        when 'cosmere' then authorized_scope(Character.all).cosmere
         else Character.none
         end
       end