+++ app/decorators_v2/cosmere_decorator.rb
@@ -7,7 +7,7 @@ class CosmereDecorator < ApplicationDecoratorV2
   ONLY_ADD_MODIFIERS = %w[str spd int wil awa pre].freeze
   WEAPON_MODIFIERS = %w[attack melee_attacks range_attacks damage melee_damage range_damage].freeze
 
-  def call(character:, simple: false, version: nil)
+  def call(character:, simple: false, version: nil) # rubocop: disable Metrics/AbcSize
     @character = character
     @version = version
     @result = character.data.attributes
@@ -24,6 +24,7 @@ class CosmereDecorator < ApplicationDecoratorV2
     @result['movement'] = modify_by_armor(movement)
     @result['features'] = apply_features
     @result['singer_forms'] = find_available_singer_forms
+    @result['investiture_max'] = @investiture ? (2 + [modified_abilities['awa'], modified_abilities['pre']].max) : 0
 
     self
   end
@@ -49,7 +50,6 @@ class CosmereDecorator < ApplicationDecoratorV2
     @result['deflect'] = equiped_armor&.dig(:items_info, 'deflect').to_i
     @result['health_max'] = health_max
     @result['focus_max'] = 2 + modified_abilities['wil']
-    @result['investiture_max'] = 2 + [modified_abilities['awa'], modified_abilities['pre']].max
     @result['load'] = find_load
     @result['movement'] = find_movement
     @result['recovery_die'] = find_recovery_die
@@ -109,7 +109,8 @@ class CosmereDecorator < ApplicationDecoratorV2
     available_features.filter_map { |feature| feature_payload(feature).merge(used_count: feature.used_count) }
   end
 
-  def feature_payload(feature)
+  def feature_payload(feature) # rubocop: disable Metrics/AbcSize
+    allow_investiture if feature.feat.info['investiture']
     {
       id: feature.id,
       slug: feature.feat.slug || feature.id,
@@ -119,7 +120,9 @@ class CosmereDecorator < ApplicationDecoratorV2
       origin: feature.feat.origin,
       origin_value: feature.feat.origin_value,
       price: feature.feat.price,
-      info: feature.feat.info
+      info: feature.feat.info,
+      continious: feature.feat.continious,
+      active: feature.active
     }.compact
   end
 
@@ -210,8 +213,17 @@ class CosmereDecorator < ApplicationDecoratorV2
         %w[discipline wil], %w[intimidation wil], %w[lore int], %w[medicine int],
         %w[deception pre], %w[insight awa], %w[leadership pre], %w[perception awa],
         %w[persuation pre], %w[survival awa]
-      ] + additional_skills.map { |(key, values)| [key, values['ability']] }
+      ] +
+        additional_skills.map { |(key, values)| [key, values['ability']] } +
+        surge_skills
     ).map { |item| skill_payload(item[0], item[1]) }
+  end
+
+  def surge_skills
+    keys = selected_skills.keys
+    [
+      %w[abrasion spd]
+    ].select { |item| keys.include?(item[0]) }
   end
 
   def skill_payload(slug, ability)
@@ -378,5 +390,9 @@ class CosmereDecorator < ApplicationDecoratorV2
       value['description'] = markdown.call(value: translate(value['description']), version: '0.4.9')
       value.slice('name', 'description')
     end.compact
+  end
+
+  def allow_investiture
+    @investiture = true
   end
 end